+++ resources/views/livewire/keuangan/jenis-anggaran.blade.php
@@ -36,12 +36,17 @@
                                             <td>{{ $counter++ }}</td>
                                             <td class="text-left">{{ $row->nama_jenis }}</td>
                                             <td class="tw-whitespace-nowrap">
-                                                <button wire:click.prevent="edit({{ $row->id }})" class="btn btn-primary" data-toggle="modal" data-target="#formDataModal">
-                                                    <i class="fas fa-edit"></i>
-                                                </button>
-                                                <button wire:click.prevent="deleteConfirm({{ $row->id }})" class="btn btn-danger">
-                                                    <i class="fas fa-trash"></i>
-                                                </button>
+                                                @if ($this->can("jenis_anggaran.edit"))
+                                                    <button wire:click.prevent="edit({{ $row->id }})" class="btn btn-primary" data-toggle="modal" data-target="#formDataModal">
+                                                        <i class="fas fa-edit"></i>
+                                                    </button>
+                                                @endif
+
+                                                @if ($this->can("jenis_anggaran.delete"))
+                                                    <button wire:click.prevent="deleteConfirm({{ $row->id }})" class="btn btn-danger">
+                                                        <i class="fas fa-trash"></i>
+                                                    </button>
+                                                @endif
                                             </td>
                                         </tr>
                                     @endforeach
@@ -56,10 +61,11 @@
                 </div>
             </div>
         </div>
-
-        <button wire:click.prevent="isEditingMode(false)" class="btn-modal" data-toggle="modal" data-backdrop="static" data-keyboard="false" data-target="#formDataModal">
-            <i class="far fa-plus"></i>
-        </button>
+        @if ($this->can("jenis_anggaran.create"))
+            <button wire:click.prevent="isEditingMode(false)" class="btn-modal" data-toggle="modal" data-backdrop="static" data-keyboard="false" data-target="#formDataModal">
+                <i class="far fa-plus"></i>
+            </button>
+        @endif
     </section>
 
     <!-- Modal Form -->
